refactor: application, stepPost 전역상태 훅 도메인별로 따로 분리
