+++ src/hooks/application/create/useHandleCreate.ts
@@ -1,7 +1,7 @@
 import { router } from "expo-router";
 
 import { useUploadImage } from "@/src/hooks/application/create/useUploadImage";
-import { useApplicationSlice } from "@/src/stores/slices/useApplicationSlice";
+import { useApplicationSlice } from "@/src/stores/slices/application/useApplicationSlice";
 import { useApplicationFormValidation } from "./useApplicationFormValidation";
 import { useCreateApplication } from "./useCreateApplication";
