+++ src/hooks/stepPost/useWorkingTimeAndWorkList.ts
@@ -1,4 +1,4 @@
-import { initialWorkingTimeAndWorkData } from "@/src/stores/slices/useApplicationSlice";
+import { initialWorkingTimeAndWorkData } from "@/src/stores/slices/stepPost/useStepPostSlice";
 import { IWorkingTimeAndWork } from "@/src/types/models/application/StepPostData";
 
 export function useWorkingTimeAndWork(
